+++ notes/models.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from django.db import models
 from django.contrib.contenttypes.models import ContentType
-from django.contrib.contenttypes import generic
+from django.contrib.contenttypes.fields import GenericForeignKey
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import ugettext_lazy as _
 from django.utils import timezone
@@ -47,7 +47,7 @@ class Note(models.Model):
     author = models.ForeignKey(user_model_label, blank=True, null=True)
     content_type = models.ForeignKey(ContentType)
     object_id = models.CharField(_('Object ID'), max_length=255)
-    content_object = generic.GenericForeignKey()
+    content_object = GenericForeignKey()
 
     created = models.DateTimeField(_('Created'), editable=False)
     modified = models.DateTimeField(_('Modified'), editable=False)